Stub sk/dark on initialization of emacs frame
This is being added as a commented piece of code because  for the
longest time I couldn't figure out why simply adding  sk/dark in
(add-hook 'after-init-hook 'sk/dark)... Found out that this line in
init-themes.el, (add-hook 'after-init-hook 'reapply-themes), was
reloading themes in custom.el and mucking up powerline. Anyways, this
is just for the record... in case I'm trying to figure this out again.